requirements.txt
Django==4.0.3
django-cors-headers==3.11.0
django-extensions==3.1.5
mysqlclient==2.1.0
bcrypt==3.2.0
PyJWT==2.3.0